Add equality comparer parameter to DeepEquals
